WIP:  added  check  `request.head()`
A light check for page existence.
    - If fails, write it down
    - Only looking for Error 404, since other kinds of codes other than 200 (like 301 - Permanently Moved) seemed to work.
    - Follow-up question: Are there other codes besides 404 that indicate dead links?
    - Follow-up question: Do we want to flag other codes (like 301) that work, but aren't in the 200 range in some way?

TODO: The request timed out so we should handle exception for time outs and other process failures.
TODO:  add some kind of instructions for testing. Since this is a highly iterative excercise, maybe just put in a `todo` for the next step and an example of the last working step.

Co-authored-by: Kishan Patel <kishanpatel789@gmail.com>
Co-authored-by: Anschel Burk <anschelburk@users.noreply.github.com>
Co-authored-by: Rana Khalil <rankhali+cisco@cisco.com>
Co-authored-by: Bob Belderbos <bob@pybit.es>